style register page
